Refactor designsafe-related types and hooks into own files
