+++ react/src/hooks/designsafe/index.ts
@@ -1,0 +1,1 @@
+export * from './useDesignSafeProjects';
