+++ react/src/hooks/index.ts
@@ -9,3 +9,4 @@ export * from './map';
 export * from './pointClouds';
 export * from './notifications';
 export * from './streetview';
+export * from './designsafe';
