+++ react/src/hooks/projects/index.ts
@@ -1,13 +1,3 @@
-export {
-  useDeleteProject,
-  useProjectsWithDesignSafeInformation,
-  useProjects,
-  useProject,
-  useDesignSafeProject,
-  useDesignSafeProjects,
-  useUpdateProjectInfo,
-} from './useProjects';
-
+export * from './useProjects';
 export { default as useCreateProject } from './useCreateProject';
-
 export { useProjectUsers } from './useProjectUsers';
