+++ react/src/hooks/projects/useProjects.ts
@@ -1,12 +1,7 @@
 import { UseQueryResult, useQueryClient } from '@tanstack/react-query';
 import { useMemo } from 'react';
-import {
-  Project,
-  DesignSafeProject,
-  DesignSafeProjectCollection,
-  ApiService,
-  ProjectRequest,
-} from '@hazmapper/types';
+import { Project, ApiService, ProjectRequest } from '@hazmapper/types';
+import { useDesignSafeProjects } from '@hazmapper/hooks/designsafe/useDesignSafeProjects';
 import { useGet, useDelete, usePut } from '@hazmapper/requests';
 import { useParams } from 'react-router-dom';
 
@@ -33,39 +28,6 @@ export const useProject = ({
     key: ['project', projectUUID],
     options,
     transform: (data) => data[0], // result is a list with a single Project
-  });
-  return query;
-};
-
-interface UseDesignSafeProjectParams {
-  designSafeProjectUUID: string;
-  options?: object;
-}
-
-export const useDesignSafeProject = ({
-  designSafeProjectUUID,
-  options = {},
-}: UseDesignSafeProjectParams): UseQueryResult<
-  DesignSafeProject | undefined
-> => {
-  const query = useGet<DesignSafeProject>({
-    endpoint: `/api/projects/v2/${designSafeProjectUUID}/`,
-    key: ['designsafe-single-projectv2', designSafeProjectUUID],
-    options,
-    apiService: ApiService.DesignSafe,
-    transform: (data) => data.baseProject,
-  });
-  return query;
-};
-
-export const useDesignSafeProjects = (): UseQueryResult<
-  DesignSafeProjectCollection | undefined
-> => {
-  // TODO add offset and a high limit (as default is 100 for this endpoint)
-  const query = useGet<DesignSafeProjectCollection>({
-    endpoint: `/api/projects/v2/`,
-    key: ['designsafe-projectsv2'],
-    apiService: ApiService.DesignSafe,
   });
   return query;
 };
