+++ react/src/types/projectsDesignSafe.ts
@@ -1,0 +1,80 @@
+interface DesignSafeProjectValue {
+  dois: string[] /* always empty, deprecated? */;
+  title: string;
+  users: any[];
+  keywords: string[];
+  nhEvents: string[];
+  dataTypes: string[];
+  projectId: string;
+  nhLatitude: string;
+  nhLocation: string;
+  description: string;
+  nhLongitude: string;
+  projectType: string;
+  hazmapperMaps?: any[];
+}
+
+/**
+ * Partial interface for DesignSafe project value data.
+ * Note: This represents key fields only, not all available fields.
+ * Some fields (e.g. dois) may be deprecated.
+ */
+interface DesignSafeProjectValue {
+  // Core project identifiers
+  projectId: string;
+  title: string;
+  description: string;
+  projectType: string;
+
+  // Users and access
+  users: any[];
+
+  // Users - Deprecated: no longer used in v2 DS
+  pi?: string;
+  coPis?: any[];
+  teamMembers?: string[];
+  guestMembers?: string[];
+
+  // Location data
+  nhLatitude: string;
+  nhLongitude: string;
+  nhLocation: string;
+
+  // Classifications and metadata
+  keywords: string[];
+  dataTypes: string[];
+  nhEvents: string[];
+  dois: string[]; // Deprecated: appears to always empty array
+
+  // Natural Hazard specific
+  nhEvent: string;
+  nhTypes: string[];
+
+  // Related features
+  hazmapperMaps?: any[];
+  facilities: string[];
+  associatedProjects: string[];
+  referencedData: any[];
+
+  // Other
+  ef?: string;
+  authors?: any[];
+  frTypes?: any[];
+  fileObjs?: any[];
+  fileTags?: string[];
+  awardNumbers?: string[];
+  tombstone?: boolean;
+}
+
+export interface DesignSafeProject {
+  uuid: string;
+  name: string;
+  value: DesignSafeProjectValue;
+  created: string;
+  lastUpdated: string;
+  associationIds?: any[];
+}
+
+export interface DesignSafeProjectCollection {
+  result?: DesignSafeProject[];
+}
